# api/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Pharma API
  version: 1.0.0
  description: API multiempresa para operação fiscal, estoque, vendas e gestão do SaaS.
servers:
  - url: http://localhost:3333
    description: Desenvolvimento
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Autentica um usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "200": { description: Tokens, usuário e empresas disponíveis }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotaciona a sessão
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200": { description: Novo par de tokens }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/auth/me:
    get:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Usuário autenticado e empresas }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/cadastros/categorias:
    get:
      tags: [Cadastros]
      security: [{ bearerAuth: [], companyHeader: [] }]
      responses:
        "200": { description: Categorias e regras fiscais da empresa }
    post:
      tags: [Cadastros]
      security: [{ bearerAuth: [], companyHeader: [] }]
      responses:
        "201": { description: Categoria criada }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/cadastros/categorias/{id}:
    put:
      tags: [Cadastros]
      security: [{ bearerAuth: [], companyHeader: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Categoria atualizada e auditada }
  /api/v1/cadastros/produtos:
    get:
      tags: [Cadastros]
      security: [{ bearerAuth: [], companyHeader: [] }]
      responses:
        "200": { description: Produtos, categorias e lotes }
    post:
      tags: [Cadastros]
      security: [{ bearerAuth: [], companyHeader: [] }]
      responses:
        "201": { description: Produto e lote inicial criados }
  /api/v1/cadastros/produtos/{id}:
    put:
      tags: [Cadastros]
      security: [{ bearerAuth: [], companyHeader: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Produto atualizado e auditado }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    companyHeader:
      type: apiKey
      in: header
      name: x-company-id
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Corpo ou parâmetro inválido
    Unauthorized:
      description: Credencial ausente, inválida ou expirada
...